Added two wavelength-shifting reflectors to the baseline geometry: one wrapped around the reentrance tube, and one along the inside of the cryostat wall. This WLSR has the reflectivity properties of Tetratex, but the WLS properties of TPB. The WLSR is only added to the geometry if the option to simulate optics is switched 'on'.
Added the option for xenon-doped LAr to the optical simulations. If XeConc is set to a value above 0, the LAr's properties are changed to resemble the properties of LAr doped with 100 ppm xenon. If there is further interest for this feature, more options for xenon concentration can be added.

Added a third-party package (CADMesh.hh) which allows the module to read .stl CAD files and incorporate the geometries into the simulation.

Added 'time' information to the Ar41 de-excitation gamma generator.

Fixed the optical map photon generator so that it effectively employs the 'confine' command available in the Geant4 GPS, to guarantee all photons are generated in the liquid argon volume.

Added settings to the optical map photon generator for producing photons in the 'inner' AAr volume, between the shield and the reentrance tube.

QOL: added a timer that reports the time taken for the simulation to process.